requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
Flask==3.0.3
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
lxml==5.3.0
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.2
pandas==2.2.1
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.1
requests==2.32.3
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==75.6.0
six==1.16.0
threadpoolctl==3.4.0
tzdata==2024.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.1.2

gunicorn